backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-bcrypt==1.0.1
pdfplumber==0.11.4
google-genai==1.0.0